+++ lib/responsive_builder.dart
@@ -3,6 +3,7 @@ library responsive_builder;
 export 'src/device_screen_type.dart';
 export 'src/helpers.dart';
 export 'src/responsive_sizing_config.dart';
+export 'src/responsive_wrapper.dart';
 export 'src/scroll/scroll_transform_item.dart';
 export 'src/scroll/scroll_transform_view.dart';
 export 'src/sizing_information.dart';